Removed background from ClojureBridge London banner graphic
Feedback on the logo design said it would be much improved without the green
backgound.

After removing the background and changing the text colour to green, the banner
feels much cleaner and blends into the page smoothly.

Including the png and orginal svg source formats for the image.  The png image
is used on page.