Updated: website style and style related plugins
Some of the content did not stand out very well, especially inline code and code
blocks. Added website.css styles to make the code stand out.

Added the wide-page plugin to spread the content wider on the page, so it looks
better on a higher resolution monitor.  The plugin has a maximum width of 95%
which was a bit high, but it was easy to over-ride this by adding a style to the
website.css file.  A width of 72% looks good and will seek further feedback on
this.