Atom and ProtoREPL editor guide
Providing an embedded view of the YouTube video I created a couple of days ago.

Described how to start a Clojure REPL for ProtoREPL, the most common ways to
evaluate code and changing to a different namespace.

Keybindings for ProtoREPL are provided as a quick reference.

Added images for MacOSX install, missing from the deployment guide commits.