Calva user guide - further improvements
Starting a project more generic now, by including Clojure CLI as a suggested
REPL selection.

Directed troubleshooting direction to starting a repl in a terminal at the
bottom of the page.

Renamed starting a remote repl to start a repl on the command line

Removed TODO from the user guide index page, as the user guide is complete